update ledger_getPoWDifficulty method
