+++ rpcapi/api/tx.go
@@ -5,7 +5,6 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"github.com/vitelabs/go-vite/chain"
 	"github.com/vitelabs/go-vite/vm/contracts/dex"
 	"math/big"
 	"math/rand"
@@ -19,10 +18,6 @@ import (
 	"github.com/vitelabs/go-vite/net"
 	"github.com/vitelabs/go-vite/verifier"
 	"github.com/vitelabs/go-vite/vite"
-	"github.com/vitelabs/go-vite/vm"
-	"github.com/vitelabs/go-vite/vm/quota"
-	"github.com/vitelabs/go-vite/vm/util"
-	"github.com/vitelabs/go-vite/vm_db"
 	"go.uber.org/atomic"
 )
 
@@ -201,17 +196,6 @@ type CalcPoWDifficultyParam struct {
 	Multiple uint16 `json:"multipleOnCongestion"`
 }
 
-type GetPoWDifficultyParam struct {
-	SelfAddr  types.Address  `json:"address"`
-	PrevHash  types.Hash     `json:"prevHash"`
-	BlockType byte           `json:"blockType"`
-	ToAddr    *types.Address `json:"toAddress"`
-	Data      []byte         `json:"data"`
-	Multiple  uint16         `json:"congestionMultiplier"`
-}
-
-var multipleDivision = big.NewInt(10)
-
 type CalcPoWDifficultyResult struct {
 	Quota         uint64  `json:"quota"`
 	QuotaRequired string  `json:"quotaRequired"`
@@ -225,88 +209,11 @@ type CalcPoWDifficultyResult struct {
 func (t Tx) CalcPoWDifficulty(param CalcPoWDifficultyParam) (result *CalcPoWDifficultyResult, err error) {
 	return calcPoWDifficulty(t.vite.Chain(), param)
 }
-func (t LedgerApi) GetPoWDifficulty(param GetPoWDifficultyParam) (result *CalcPoWDifficultyResult, err error) {
-	return calcPoWDifficulty(t.chain,
-		CalcPoWDifficultyParam{param.SelfAddr, param.PrevHash, param.BlockType, param.ToAddr, param.Data, true, param.Multiple})
-}
-func calcPoWDifficulty(c chain.Chain, param CalcPoWDifficultyParam) (result *CalcPoWDifficultyResult, err error) {
-	latestBlock, err := c.GetLatestAccountBlock(param.SelfAddr)
-	if err != nil {
-		return nil, err
-	}
-	if (latestBlock == nil && !param.PrevHash.IsZero()) ||
-		(latestBlock != nil && latestBlock.Hash != param.PrevHash) {
-		return nil, util.ErrChainForked
-	}
-	// get quota required
-	block := &ledger.AccountBlock{
-		BlockType:      param.BlockType,
-		AccountAddress: param.SelfAddr,
-		PrevHash:       param.PrevHash,
-		Data:           param.Data,
-	}
-	if param.ToAddr != nil {
-		block.ToAddress = *param.ToAddr
-	} else if param.BlockType == ledger.BlockTypeSendCall {
-		return nil, errors.New("toAddress is nil")
-	}
-	sb := c.GetLatestSnapshotBlock()
-	db, err := vm_db.NewVmDb(c, &param.SelfAddr, &sb.Hash, &param.PrevHash)
-	if err != nil {
-		return nil, err
-	}
-	quotaRequired, err := vm.GasRequiredForBlock(db, block, util.GasTableByHeight(sb.Height), sb.Height)
-	if err != nil {
-		return nil, err
-	}
-
-	qc, _, isCongestion := quota.CalcQc(db, sb.Height)
-
-	// get current quota
-	var pledgeAmount *big.Int
-	var q types.Quota
-	if param.UsePledgeQuota {
-		pledgeAmount, err = c.GetPledgeBeneficialAmount(param.SelfAddr)
-		if err != nil {
-			return nil, err
-		}
-		q, err := quota.GetPledgeQuota(db, param.SelfAddr, pledgeAmount, sb.Height)
-		if err != nil {
-			return nil, err
-		}
-		if q.Current() >= quotaRequired {
-			return &CalcPoWDifficultyResult{quotaRequired, Uint64ToString(quotaRequired), "", Float64ToString(float64(quotaRequired)/float64(quota.QuotaForUtps), 4), bigIntToString(qc), isCongestion}, nil
-		}
-	} else {
-		pledgeAmount = big.NewInt(0)
-		q = types.NewQuota(0, 0, 0, 0, false)
-	}
-	// calc difficulty if current quota is not enough
-	canPoW := quota.CanPoW(db, block.AccountAddress)
-	if !canPoW {
-		return nil, util.ErrCalcPoWTwice
-	}
-	d, err := quota.CalcPoWDifficulty(db, quotaRequired, q, sb.Height)
-	if err != nil {
-		return nil, err
-	}
-	if isCongestion && param.Multiple > uint16(multipleDivision.Uint64()) {
-		d.Mul(d, multipleDivision)
-		d.Div(d, big.NewInt(int64(param.Multiple)))
-	}
-	return &CalcPoWDifficultyResult{quotaRequired, Uint64ToString(quotaRequired), d.String(), Float64ToString(float64(quotaRequired)/float64(quota.QuotaForUtps), 4), bigIntToString(qc), isCongestion}, nil
-}
 
 type CalcQuotaRequiredParam struct {
 	SelfAddr  types.Address  `json:"selfAddr"`
 	BlockType byte           `json:"blockType"`
 	ToAddr    *types.Address `json:"toAddr"`
-	Data      []byte         `json:"data"`
-}
-type GetQuotaRequiredParam struct {
-	SelfAddr  types.Address  `json:"address"`
-	BlockType byte           `json:"blockType"`
-	ToAddr    *types.Address `json:"toAddress"`
 	Data      []byte         `json:"data"`
 }
 type CalcQuotaRequiredResult struct {
@@ -317,41 +224,6 @@ type CalcQuotaRequiredResult struct {
 // Deprecated: use ledger_getRequiredQuota instead
 func (t Tx) CalcQuotaRequired(param CalcQuotaRequiredParam) (*CalcQuotaRequiredResult, error) {
 	return calcQuotaRequired(t.vite.Chain(), param)
-}
-func (t LedgerApi) GetRequiredQuota(param GetQuotaRequiredParam) (*CalcQuotaRequiredResult, error) {
-	return calcQuotaRequired(t.chain,
-		CalcQuotaRequiredParam{param.SelfAddr, param.BlockType, param.ToAddr, param.Data})
-}
-func calcQuotaRequired(c chain.Chain, param CalcQuotaRequiredParam) (*CalcQuotaRequiredResult, error) {
-	latestBlock, err := c.GetLatestAccountBlock(param.SelfAddr)
-	if err != nil {
-		return nil, err
-	}
-	prevHash := types.Hash{}
-	if latestBlock != nil {
-		prevHash = latestBlock.Hash
-	}
-	// get quota required
-	block := &ledger.AccountBlock{
-		BlockType:      param.BlockType,
-		AccountAddress: param.SelfAddr,
-		Data:           param.Data,
-	}
-	if param.ToAddr != nil {
-		block.ToAddress = *param.ToAddr
-	} else if param.BlockType == ledger.BlockTypeSendCall {
-		return nil, errors.New("toAddress is nil")
-	}
-	sb := c.GetLatestSnapshotBlock()
-	db, err := vm_db.NewVmDb(c, &param.SelfAddr, &sb.Hash, &prevHash)
-	if err != nil {
-		return nil, err
-	}
-	quotaRequired, err := vm.GasRequiredForBlock(db, block, util.GasTableByHeight(sb.Height), sb.Height)
-	if err != nil {
-		return nil, err
-	}
-	return &CalcQuotaRequiredResult{Uint64ToString(quotaRequired), Float64ToString(float64(quotaRequired)/float64(quota.QuotaForUtps), 4)}, nil
 }
 
 func (tx Tx) autoSend() {
